* commit 'cd9608ccc651e7554817f329afdc27181fa79ab9':
  Changed name of the new function from checkAndResize to checkResizeClear following Simons comment.
  Add a function for resizing of vectors with high capacity.
  Only one comment on the resizing procedure, and higher limits for resizing.
  Added the vector resizing to the PXDCKF, and also added comments why this is done in the way it's implemented.
  Replaced double by float where using the stateCache in the PXD CKF.
  Replaced double by float when using the CKFState stateCache for SVD CKF.
  Optimized layout of the CKF state to reduce memory footprint, and introduced feature to start with an empty vector in the SVDCKF in case the current vector is large and potentially causes large unused memory footprint.

+++ tracking/ckf/pxd/filters/relations/include/SensorPXDPairFilter.h
@@ -23,8 +23,8 @@ namespace Belle2 {
 
   private:
     /// Pre-filter relations in phi between seed states and hit states
-    double m_param_PhiRecoTrackToHitCut = 0.8;
+    float m_param_PhiRecoTrackToHitCut = 0.8;
     /// Pre-filter relations in theta between hit states
-    double m_param_PhiHitHitCut = 1.0;
+    float m_param_PhiHitHitCut = 1.0;
   };
 }
