+++ tracking/ckf/svd/filters/relations/include/LooseSVDPairFilter.h
@@ -23,12 +23,12 @@ namespace Belle2 {
 
   private:
     /// Pre-filter relations in theta for overlay
-    double m_param_ThetaOverlayRegionPrecut = 0.4;
+    float m_param_ThetaOverlayRegionPrecut = 0.4;
     /// Pre-filter relations in phi between seed states and hit states
-    double m_param_PhiSeedHitPrecut = M_PI / 3.;
+    float m_param_PhiSeedHitPrecut = M_PI / 3.;
     /// Pre-filter relations in phi between hit states
-    double m_param_PhiHitHitPrecut = M_PI / 3.;
+    float m_param_PhiHitHitPrecut = M_PI / 3.;
     /// Pre-filter relations in theta between hit states
-    double m_param_ThetaHitHitPrecut = M_PI / 3.;
+    float m_param_ThetaHitHitPrecut = M_PI / 3.;
   };
 }
