+++ tracking/ckf/svd/filters/relations/include/SensorSVDPairFilter.h
@@ -23,8 +23,8 @@ namespace Belle2 {
 
   private:
     /// Pre-filter relations in phi between seed states and hit states
-    double m_param_PhiRecoTrackToHitCut = 0.5;
+    float m_param_PhiRecoTrackToHitCut = 0.5;
     /// Pre-filter relations in theta between hit states
-    double m_param_PhiHitHitCut = 1.14159265359;
+    float m_param_PhiHitHitCut = 1.14159265359;
   };
 }
