+++ tracking/ckf/svd/filters/relations/src/DistanceSVDPairFilter.cc
@@ -26,13 +26,13 @@ DistanceSVDPairFilter::operator()(const std::pair<const CKFToSVDState*, const CK
 
   B2ASSERT("You have filled the wrong states into this!", toStateCache.isHitState);
 
-  double phiDiff = fromStateCache.phi - toStateCache.phi;
+  float phiDiff = fromStateCache.phi - toStateCache.phi;
   while (phiDiff > M_PI) phiDiff -= 2. * M_PI;
   while (phiDiff < -M_PI) phiDiff += 2. * M_PI;
 
   if (not fromStateCache.isHitState) {
     // We are coming from a CDC track, so we can use its position to only look for matching ladders
-    if (fabs(phiDiff) < 0.2 and fabs(fromStateCache.theta - toStateCache.theta) < 0.2) {
+    if (fabs(phiDiff) < 0.2f and fabs(fromStateCache.theta - toStateCache.theta) < 0.2f) {
       return 1.0;
     }
     // If the current state (fromState) is a RecoTrack-based state, but no relations could be created
@@ -47,7 +47,7 @@ DistanceSVDPairFilter::operator()(const std::pair<const CKFToSVDState*, const CK
     return 1.0;
   }
 
-  if (abs(phiDiff) < 0.2) {
+  if (abs(phiDiff) < 0.2f) {
     return 1.0;
   }
